requirements.txt
azure-ai-inference
ipykernel
openai
streamlit
firebase-admin
